requirements.txt
streamlit>=1.37.0
plotly>=5.23.0
pandas>=2.0.0
numpy>=1.24.0
pydeck>=0.9.0
pyyaml>=6.0
pytest>=8.0.0
folium>=0.17.0
streamlit-folium>=0.22.0
meteostat>=2.1.0
